+++ bizyair_extras/__init__.py
@@ -13,6 +13,7 @@ from .nodes_ipadapter_plus.nodes_ipadapter_plus import *
 from .nodes_kolors_mz import *
 from .nodes_model_advanced import *
 from .nodes_sd3 import *
+from .nodes_segment_anything import *
 from .nodes_testing_utils import *
 from .nodes_ultimatesdupscale import *
 from .nodes_upscale_model import *
